putting config in init
